Fixing verison number and tests
