+++ frontend/app-development/features/SettingsModalButton/SettingsModal/components/Tabs/AbouTab/AboutTab.module.css
@@ -16,6 +16,7 @@
   display: flex;
   align-items: center;
   margin-block: 10px;
+  color: var(--fds-semantic-text-action-active);
 }
 
 .label {
@@ -29,16 +30,18 @@
 
 .createdByParagraph {
   margin-left: 10px;
-}
-
-.createdForIcon {
-  font-size: 2.5rem;
+  color: inherit;
 }
 
 .createdByIcon {
   font-size: 1.5rem;
+  color: inherit;
 }
 
 .avatar {
   max-height: 2em;
+}
+
+.dateParagraph {
+  color: var(--fds-semantic-text-neutral-subtle);
 }
